\name{ABarrayGUI}
\alias{ABarrayGUI}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ GUI for ABarray to perform QA, data transformation and statistical analysis }
\description{
  A front end GUI for ABarray package to perform data analysis.
}
\usage{
ABarrayGUI()
}
%- maybe also 'usage' for other objects documented here.
\arguments{
}

\details{
  The interface gathers required paramters for the ABarray packages to
  run. See ABarray for more details.
  }

\value{
  No return values.
}

\author{ Yongming Sun <sunya@appliedbiosystems.com> }

\seealso{ ABarray, doPlotEset, doPlotFCT, doANOVA, matrixPlot, mvaPair2, doLPE,
  doVennDiagram, hclusterPlot }

\examples{
#- ABarrayGUI()
}
\keyword{ manip }% at least one, from doc/KEYWORDS
