\name{cvv}
\alias{cvv}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ CV calculation }
\description{
  Calculate cv 
}
\usage{
cvv(data)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{data}{ data matrix contain expression values }
}

\value{
  vector of cv for each gene or probe
}

\author{ Yongming Sun }

\examples{
%-cvVector <- cvv(exprMatrix)
}
\keyword{ manip }% at least one, from doc/KEYWORDS

