\name{cvvPlot}
\alias{cvvPlot}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Plot CV value }
\description{
  Plot CV value against average intensity (log2)
}
\usage{
cvvPlot(data, name)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{data}{ vector of cv for each gene }
  \item{name}{ name of the plot }
}

\value{
  None
}

\author{ Yongming Sun }

\seealso{ cvv \code{\link{cvv}} to calulate cv }
\examples{
%-cvvPlot(cvData, "plotTitle")
}
\keyword{ hplot }% at least one, from doc/KEYWORDS

