\name{panel.cor}
\alias{panel.cor}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Create correlation panel }
\description{
  Create correlation panel
}
\usage{
panel.cor(x, y, digits=3, prefix="", cex.cor)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ vector of expression value for one sample }
  \item{y}{ vector of expression value for another sample }
  \item{digits}{ number of digits to display the correlation }
  \item{prefix}{ additional text to display }
  \item{cex.cor}{ size of the text }
}

\value{
  None
}
\author{ Yongming Sun }

\examples{
##---- Not intended for direct function call !! ----
}
\keyword{ hplot }% at least one, from doc/KEYWORDS

