\name{panel.scatter}
\alias{panel.scatter}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Creat scatter plot }
\description{
  Create scatter plot
}
\usage{
panel.scatter(x, y, col = "blue", bg = NA, pch = ".", 
    cex = 1, col.smooth = "red", span = 2/3, iter = 3, ...) 
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{ vector of expression for one sample }
  \item{y}{ vector of expression for another sample }
  \item{col}{ color of points }
  \item{bg}{ background colors }
  \item{pch}{ pch paremeter }
  \item{cex}{ size of text }
  \item{col.smooth}{ color of smooth line }
  \item{span}{ span of the plot }
  \item{iter}{ iteration }
  \item{...}{ additional arguments }
}

\value{
  None
}

\author{ Yongming Sun }

\examples{
##---- Not intended for use this function directly !! ----
}
\keyword{ hplot }% at least one, from doc/KEYWORDS

