\name{rgcolorsfunc}
\alias{rgcolorsfunc}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ generate color map }
\description{
  Generate color map for heatmap use
}
\usage{
rgcolorsfunc(n = 50)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{n}{ number of colors to generate }
}

\value{
  rgb color vector
}

\author{ Yongming Sun }

\examples{
 ## Do not call this function directly
 rgb <- rgcolorsfunc()
}
\keyword{ manip }% at least one, from doc/KEYWORDS

